fw_fcp: improve AV/C interim response handling
In AV/C Digital Interface Command Set General Specification Version 4.3,
there's a 'interim' type of response transaction. When transaction
originator receives this type of response transaction, it should wait
for later transaction to receive meaningful response.

Current implementation doesn't follow to this behaviour because of a bug
of usage of GArray type. This commit fixes this bug.