fw_req: fix to set initial value to FW_REQ_PROP_TYPE_TIMEOUT
A commit 91bc3572ba64 ("fw_req: timeout as an object property") adds
timeout property to HinawaFwReq object. This property has no attributes so
that it has non-zero value as its initial state. As a result, transactions
often get timeout for models which return responses over a period of task
scheduling. The period is variant and this does not always occur.

This commit fixes the bug.

Fixes: 91bc3572ba64 ("fw_req: timeout as an object property")
Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>