fw_resp: handle guint8 array instead of guint32 array
In libhinawa v0.x.0, HinawaFwResp GObject class handles data in a shape
of guint32 array for consumers of exported API. However, IEEE 1222
defines transaction which is not for quadlet data. For this type of
transaction, current design of the gobject class is not suitable.

This commit changes the gobject class to handle guint8 array. This loses
backward compatibility of libhinawa0.