fw_unit: handle GET_INFO request properly in a point of config_rom
On IEEE 1394 bus, it's likely that each node changes contents of
its configuration ROM. In this case, the length of config_rom can be
different. For this design, Linux firewire subsystem takes GET_INFO
ioctl request in account. Userland applications call the request
with buffer in userspace and its length, then Linux firewire subsystem
fills the buffer with actual contents of configuration ROM, up to 256
quadlets.

This commit keeps 1024 bytes (=256 quadlets) for configuration ROM in
userspace for a case that actual length of configuration ROM changes.