fw_resp: change prototype of handler for "requested" signal
HinawaFwResp has "requested" GObject signal. A receiver is expected to
handle requested buffer and returns response buffer. However, a design
fault of PyGObject, whole content of the response buffer is not passed
to event emitter if the buffer is passed as an argument to the signal
handler.

This commit adds two methods to get requested buffer and set response
buffer to HinawaFwResp GObject object. The receiver is expected to call
these methods in its handler for the GObject signal, then returns rcode of
IEEE 1394 transaction.