meson: attempt to build with meson
Meson is a frontend of buid system with several backend. In Linux
environment, Ninja is the backend. The combination of Meson and
Ninja is designed for its build speed. As long as I know, the
combination is faster than the combination of GNU Autotools/Make
by double. Additionally, several projects in GNOME and freedesktop.org
are already adopted to meson, thus libhinawa get some advantages
because it's an application of GLib.

This commit adopts to Meson.