bump up to v1.0.0
Since v.0.9.0, this library loses backward compatibility for below
points:
 * HinawaSndUnit loses some methods to delegate tasks to HinawaFwReq.
 * HinawaFwResp changes procedure of signal handling due to arguments of
   data array.
 * HinawaFwReq changes prototype of methods to handle byte-aligned buffer.
 * HinawaFwUnit changes prototype of a method to handle byte-aligned
   buffer.

Additionally, interface versioning scheme includes a mistake. This can
brings backward number of interface version at package v1.0.0 release.
This mistake is fixed in a previous commit and I need to bump package
version up to v1.0.0. As a result, this library loses backward
compatibility and gir version is updated to 2.0 as well.