fw_resp: fix failure to deallication of handler to address space
I forget to keep a pointer in parameter returned from ioctl(2) with
FW_CDEV_IOC_ALLOCATE. As a result, a call of ioctl(2) with
FW_CDEV_IOC_DEALLOCATE doesn't work as expected. This influences
some invalid accesses to process' VMA after a call of
'hinawa_fw_resp_unregister()'.

This commit fixes the bug.