fw_unit/snd_unit: use 500 msec timeout for poll to block processing
This library executes poll(2) with blocking mode to save consumption of
CPU time, however there's no way to wake up from the blocking except
for emitting any UNIX signal. In a convention of design of library,
it's not good to emit signal to return from the blocking.

This commit expands timeout up to 500 msec. As a result, I/O thread
can return safely at cancellation.

Fixes: 8205f2c40ab1('fw_unit/snd_unit: use blocking poll to save CPU usage')