context: maintain two threads for I/O operation
In a typical use case of this library, it's probable to start firewire
transaction in a handler of ALSA hwdep device, according to asynchronous
events from actual device.

At present, libhinawa starts an I/O thread to handle events from both
of FireWire character device and ALSA hwdep device. This design doesn't
satisfy the above use case.

This commit maintains two threads for the devices to support the use
case.