fw_req: use 200 msec as a default timeout of transaction
In IEEE 1394a:2000 specification, the minimum value of transaction timeout
is defined to 100 msec. Although this is used for transaction initiated by
userspace application ideally, the application can be scheduled enough
later or interrupt handling is enough delayed till it's occurs. Furthermore,
IEEE 1394 defines 'split transaction' to expand time for response,
depending on transaction layer of target device.

This commit uses 200 msec as a default timeout in HinawaFwReq for safe.