fix workflow of Github Actions to clone requested repository
Even if any pull request is triggered, in 'build_in_ubuntu_on_lxd' job
of 'Build test' workflow, upstream repository is cloned and requested
git hash is checked out. However this fails because the hash is in
requested repository.

This commit fixes the workflow so that requested repository is cloned.

Fixes: 25cf45ea1582 ('Add support workflow of Github Actions')