add support for i386 build job to workflow of Github Actions
In lxd repository for Ubuntu Cloud Images[1], Ubuntu i386 container
is available for applications of Intel386 ABI.

This commit uses the container to test build for Intel386 ABI in
workflow of Github Actions.

I note that no official Fedora i386 container is in dockerhub
registry[2]. It's needed to find the way to build application for
Intel386 ABI on AMD64 architecture (perhaps multiarch) if we are
going to support it.

[1] https://cloud-images.ubuntu.com/releases/
[2] https://hub.docker.com/_/fedora