meson: build HinawaSndUnit derived GObject objects always
Some kernel features are missing in releases of Linux kernel v4.0.
For the reason, backport header is included. Since Linux kernel v5.0 or
later, the backport header is useless. The last release of v4.0 is the
end of 2018. It's worth to drop the support.

This commit is a preparation. This commit builds HinawaSndUnit derived
GObject objects always regardless of kernel version.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>