hinawa_error: add error domain for libhinawa
The libhinawa has multiple instances of GQuark for domain of errors in
each objects. Although symbols for functions to retrieve the GQuark
instances are globally exported, no API is provided for applications.
Therefore it's possible to add changes relevant to the instances without
any regressions for applications.

This commit unifies the instances into single, therefore one function
is added to retrieve the instance, hinawa_error_quark().

Instances of GError with the instance as domain have error code from
errno.h in Linux system. Although this is against GLib recommendation,
it's hard to cover all error path in kernel drivers.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>