snd_unit: error reporting enhancement
In guideline of GLib's GError, the naming scheme of GQuark for
error domain is <namespace>_<module>_error_quark. Additionally,
GLib enumeration should be implemented for error code and the naming
scheme is <NAMESPACE>_<MODULE>_ERROR_<CODE>. GError should be
instanciated just for recoverable runtime error.

Current implementation of GError for HinawaSndUnit is against the
above. This commit obsoletes current GQuark and add new one. A function
is added for applications to access to it. This commit adds enumeration
for error code as well. The error code is just used by GError instance
with domain of the GQuark.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>