fw_req: error reporting enhancement
In guideline of GLib's GError, the naming scheme of GQuark for
error domain is <namespace>_<module>_error_quark. Additionally,
GLib enumeration should be implemented for error code and the naming
scheme is <NAMESPACE>_<MODULE>_ERROR_<CODE>. GError should be
instanciated just for recoverable runtime error.

Current implementation of GError for HinawaFwFcp is against the
above. This commit obsoletes current GQuark and add new one. A function
is added for applications to access to it.

Against the guideline, this commit adds no enumeration for error code
since HinawaFwRcode enumeration already represents the error code.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>