snd_unit: improve check of combination between type of hwdep device and the type of GObject
Some GObject-derived classes are specific to unit, therefore it's not
good to instantiate the class for uncorresponding unit. Although each
classes has check to avoid the issue, it's not enough implementation
because file descriptor for ALSA hwdep character device is not closed
even if it's error case.

This commit improves the check of combination between the type of
unit associated to the hwdep device and the type of GObject class going
to be instantiated. In error case, the file descriptor is closed.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>