fw_node: process GError in ioctl method
This commit is a preparation for enhancement of error reporting.

At present, error code returns from the call of internal ioctl API
for firewire character device. This allows calles to report each error
case. On the other hand, it can bring cumbersome code to callers.
It's better to unify generation of error reporting.

This commit takes the internal API to return the instance of GError at
error case instead of error code.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>