fw_req: add 'responded' GObject signal to notify response subaction of transaction
In IEEE 1394, asynchronous transaction consists of two subactions;
request and response. Linux firewire character device allows userspace
applications to initiate the request subaction, then to read response
subaction when arrived.

Current implementation of HinawaFwReq has API to finish the transaction
synchronously. However, it's convenient for some kind of applications just
to initiate the request subaction, then wait for notification of response
subaction.

This commit is a preparation for the notification. A new 'responded'
GObject signal is added.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>