fw_node: internal code refactoring to handle content of configuration ROM
Linux FireWire subsystem caches the content of configuration ROM by host-
endian. I misunderstood it by big-endian and use GUINT32_FROM_BE() macro
to convert from big-endian to host-endian. As a result, the content is
aligned to big-endian.

This commit corrects the usage of macro with GUINT32_TO_BE(). The
alignment is still kept from applications.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>