fw_resp: add API to look up region of address and reserve range of address.
In interface version 4 or above, Linux FireWire subsystem allows
userspace applications to reserve range of address within specified region
of address. This is useful when the application requires range of
address at arbitrary position in the region.

This commit adds new API to look up within region of address and reserve
range of address.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>