meson: obsolete conditional build
The library supported conditional build since its early stage, mainly due
to update in UAPI of ALSA firewire stack. Nowadays the most of component
in the header is disseminated according to newer versions of Linux kernel.
It's time to obsolete the conditional build to take it easy for optimization
to build system.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>