remove launcher for sample scripts
The run.sh script was originally added to load build library and GIR for
Python 3 process by configure environment variables.

In methon build system, the directory to build is adjustable by hand,
therefore it's impossible to indicate the directory in advance to load
build library and GIR.

Nowadays, the script is out-of-date. Let me remove it.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>