cycle_time: add methods to compute isochronous cycle from time stamp
When comparing to isochronous cycle calculated from the value of
CYCLE_TIME register, the second field in time stamp retrieved from
each context of 1394 OHCI should be completed, since it has just
three bits.

This commit adds new method to compute the isochronous cycle from the
time stamp as well as new method to parse the time stamp as is.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>