cycle_time: add boxed object to express cycle time
It is helpful for application to read current value of CYCLE_TIME
register in 1394 OHCI controller. Linux FireWire subsystem allows
user space application to read it as well as system time according to
clock id. The read data is aggregated.

This commit adds boxed object to express the aggregated data.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>