fw_req: remove 'sync' suffix from Hinawa.FwReq.transaction_with_tstamp()
The suffixes, 'sync' and 'async', in name of method is special in GNOME
project. They are related to usage of GTask for asynchronous I/O, while
current usage in the library is for blocking I/O.

This commit removes 'sync' suffix from the added API. In future 3.0
release, existent method, hinawa_fw_req_transaction(), will be deleted
once, then it will be redefined as the one without time stamp.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>