fw_req: add Hinawa.FwReq.request() to obsolete usage of 'async'
The suffixes, 'sync' and 'async', in name of method is special in GNOME
project. They are related to usage of GTask for asynchronous I/O, while
current usage in the library is for blocking I/O.

This commit adds new method, hinawa_fw_req_request() and obsoletes usage
of 'async'.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>