fw_fcp: iterate list entries for transactions
The pointer in stack is passed to the thread in which response is copied
to it, however the lifetime of stack is different from the lifetime of
thread. It can be caused that the pointer is accessed in the thread after
the stack is dropped.

This commit solves the issue by enumerating the list of transactions.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>