fw_node: add card_id property for the numeric ID of 1394 OHCI hardware
When multiple 1394 OHCI hardware are available, Linux FireWire subsystem
maintains them by numeric index. The instance of node device has a property
about the numeric index and it is available in fw_cdev_get_info structure.

The numeric index is passed to an argument for Hinawa.FwResp::requested
signal, thus it is convenient for user space application to retrieve the
numeric index via property of Hinawa.FwNode.

This commit adds the property.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>