fw_fcp: record history of node ID for transaction peer
Hinawa.FwFcp is responsible for delivery of transaction response from
peer node to application by Hinawa.FwFcp::responded signal. However,
when several nodes connecting to the same 1394 OHCI hardware, current
implementation allows application to dispatch the signal in the response
from non-peer node. The check of generation and src_node_id parameters
of response is useful to the delivery, but the change of bus topology
is concerned since Linux FireWire subsystem has no guarantee the order
of events for bus-reset and asynchronous request for transaction
response. It is possible to dispatch the request event before handling
the bus-reset event

As a workaround, this commit adds history of node ID for transaction
peer to filter out the transaction response from non-peer node. As a
result, the signal is emitted exactly for the response transaction from
the peer node.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>